+++ src/control/server/instance_drpc_test.go
@@ -1,0 +1,170 @@
+//
+// (C) Copyright 2020 Intel Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// GOVERNMENT LICENSE RIGHTS-OPEN SOURCE SOFTWARE
+// The Government's rights to use, modify, reproduce, release, perform, display,
+// or disclose this software are subject to the terms of the Apache License as
+// provided in Contract No. 8F-30005.
+// Any reproduction of computer software, computer software documentation, or
+// portions thereof marked with this legend must also reproduce the markings.
+//
+
+package server
+
+import (
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/golang/protobuf/proto"
+	"github.com/google/go-cmp/cmp"
+	"github.com/pkg/errors"
+
+	"github.com/daos-stack/daos/src/control/common"
+	mgmtpb "github.com/daos-stack/daos/src/control/common/proto/mgmt"
+	srvpb "github.com/daos-stack/daos/src/control/common/proto/srv"
+	"github.com/daos-stack/daos/src/control/drpc"
+	"github.com/daos-stack/daos/src/control/logging"
+	. "github.com/daos-stack/daos/src/control/system"
+)
+
+func getTestNotifyReadyReq(t *testing.T, sockPath string, idx uint32) *srvpb.NotifyReadyReq {
+	return &srvpb.NotifyReadyReq{
+		DrpcListenerSock: sockPath,
+		InstanceIdx:      idx,
+	}
+}
+
+func waitForIosrvReady(t *testing.T, instance *IOServerInstance) {
+	select {
+	case <-time.After(100 * time.Millisecond):
+		t.Fatal("IO server never became ready!")
+	case <-instance.awaitDrpcReady():
+		return
+	}
+}
+
+func TestIOServerInstance_NotifyDrpcReady(t *testing.T) {
+	log, buf := logging.NewTestLogger(t.Name())
+	defer common.ShowBufferOnFailure(t, buf)
+
+	instance := getTestIOServerInstance(log)
+
+	req := getTestNotifyReadyReq(t, "/tmp/instance_test.sock", 0)
+
+	instance.NotifyDrpcReady(req)
+
+	dc, err := instance.getDrpcClient()
+	if err != nil || dc == nil {
+		t.Fatal("Expected a dRPC client connection")
+	}
+
+	waitForIosrvReady(t, instance)
+}
+
+func TestIOServerInstance_CallDrpc(t *testing.T) {
+	for name, tc := range map[string]struct {
+		notReady bool
+		resp     *drpc.Response
+		expErr   error
+	}{
+		"not ready": {
+			notReady: true,
+			expErr:   errors.New("no dRPC client set"),
+		},
+		"success": {
+			resp: &drpc.Response{},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			log, buf := logging.NewTestLogger(t.Name())
+			defer common.ShowBufferOnFailure(t, buf)
+			instance := getTestIOServerInstance(log)
+			if !tc.notReady {
+				cfg := &mockDrpcClientConfig{
+					SendMsgResponse: tc.resp,
+				}
+				instance.setDrpcClient(newMockDrpcClient(cfg))
+			}
+
+			_, err := instance.CallDrpc(drpc.ModuleMgmt, drpc.MethodPoolCreate, &mgmtpb.PoolCreateReq{})
+			common.CmpErr(t, tc.expErr, err)
+		})
+	}
+}
+
+func TestIOServerInstance_DrespToRankResult(t *testing.T) {
+	dRank := Rank(1)
+
+	for name, tc := range map[string]struct {
+		daosResp    *mgmtpb.DaosResp
+		inErr       error
+		targetState MemberState
+		junkRPC     bool
+		expResult   *MemberResult
+	}{
+		"rank success": {
+			expResult: &MemberResult{Rank: dRank, Action: "test", State: MemberStateJoined},
+		},
+		"rank failure": {
+			daosResp: &mgmtpb.DaosResp{Status: -1},
+			expResult: &MemberResult{
+				Rank: dRank, Action: "test", State: MemberStateErrored, Errored: true,
+				Msg: fmt.Sprintf("rank %d dRPC returned DER -1", dRank),
+			},
+		},
+		"drpc failure": {
+			inErr: errors.New("returned from CallDrpc"),
+			expResult: &MemberResult{
+				Rank: dRank, Action: "test", State: MemberStateErrored, Errored: true,
+				Msg: fmt.Sprintf("rank %d dRPC failed: returned from CallDrpc", dRank),
+			},
+		},
+		"unmarshal failure": {
+			junkRPC: true,
+			expResult: &MemberResult{
+				Rank: dRank, Action: "test", State: MemberStateErrored, Errored: true,
+				Msg: fmt.Sprintf("rank %d dRPC unmarshal failed: proto: mgmt.DaosResp: illegal tag 0 (wire type 0)", dRank),
+			},
+		},
+	} {
+		t.Run(name, func(t *testing.T) {
+			_, buf := logging.NewTestLogger(t.Name())
+			defer common.ShowBufferOnFailure(t, buf)
+
+			if tc.daosResp == nil {
+				tc.daosResp = &mgmtpb.DaosResp{Status: 0}
+			}
+			if tc.targetState == MemberStateUnknown {
+				tc.targetState = MemberStateJoined
+			}
+
+			// convert input DaosResp to drpcResponse to test
+			rb := makeBadBytes(42)
+			if !tc.junkRPC {
+				rb, _ = proto.Marshal(tc.daosResp)
+			}
+			resp := &drpc.Response{
+				Status: drpc.Status_SUCCESS, // this will already have been validated by CallDrpc
+				Body:   rb,
+			}
+
+			gotResult := drespToMemberResult(Rank(dRank), "test", resp, tc.inErr, tc.targetState)
+			if diff := cmp.Diff(tc.expResult, gotResult, common.DefaultCmpOpts()...); diff != "" {
+				t.Fatalf("unexpected response (-want, +got)\n%s\n", diff)
+			}
+		})
+	}
+}
